// src/app/shared/model/CommonPayload.ts
export class Common {
    title : String = "";
    discription : String ="";
    subCategoryId: Number = 0;
    categoryId: Number = 0;
    price: Number = 0;
    pincode: String = "";
    state: String = "";
    city: String = "";
    nearBy: String = "";
    name: String ="";
    mobile: String= "";
    isPremium: Boolean = false;
    isActive: Boolean = true;
    createdBy: Number= 0;
    createdOn:String= "";
    modifiedBy: Number = 0;
    modifiedOn: String = "";
}